requirements.txt
azure-iot-device
azure-iot-hub

